+++ gt_manage/miniprogram/pages/PagePer/deliveryOrder/deliveryOrder.js
@@ -1,19 +1,51 @@
 // miniprogram/pages/PagePer/devliveryOrder/devliveryOrder.js
+import {
+  CloudFn
+} from '../../../utils/CloudFn'
+const cloudFn = new CloudFn()
 Page({
 
   /**
    * 页面的初始数据
    */
   data: {
-
+    deliveryArray: [], // 待配送列表
   },
 
   /**
    * 生命周期函数--监听页面加载
    */
   onLoad: function (options) {
-
+   
   },
+
+  // 获取全部数据
+  _loadData: function () {
+    wx.showLoading({
+      title: '加载中...',
+      icon: 'none'
+    })
+    cloudFn.$callFn({
+      data: {
+        name: 'CloudAPIBase',
+        body: {
+          fn: 'get',
+          base: 'user-order',
+          is_where: false,
+          where_data: {
+            state: 2
+          }
+        }
+      }
+    }).then(res => {
+      wx.hideLoading()
+      this.data.deliveryArray = res.obj
+      this.setData({
+        deliveryArray: this.data.deliveryArray
+      })
+    })
+  },
+
 
   /**
    * 生命周期函数--监听页面初次渲染完成
@@ -26,7 +58,7 @@ Page({
    * 生命周期函数--监听页面显示
    */
   onShow: function () {
-
+    this._loadData()
   },
 
   /**